Plugin: SQLite3 updated to 3.33.0 using new project
Change the project Sqleet to SQLite3 Multiple Ciphers to support
encryption extensions. The project SQLite3 Multiple Ciphers implements
an encryption extension for SQLite with support for multiple ciphers.
This project is the successor of Sqleet with more encryption cipher
schemes supported.

Web pages of the project: https://github.com/utelle/SQLite3MultipleCiphers

The current implementation of the SQLite3 Multiple Ciphers in the U++
is backward compatible with Sqleet.